# app/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "iSECTECH Notification & Trust Score API",
    "description": "Production-grade API for notification management and trust scoring systems supporting mobile PWA and enterprise security operations",
    "version": "1.0.0",
    "contact": {
      "name": "iSECTECH API Team",
      "email": "api@isectech.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://isectech.com/license"
    }
  },
  "servers": [
    {
      "url": "https://api.isectech.com/v1",
      "description": "Production server"
    },
    {
      "url": "https://api-staging.isectech.com/v1",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/notifications/templates/render": {
      "post": {
        "tags": ["Notification Templates"],
        "summary": "Render notification template",
        "description": "Render a template with variables and personalization",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template rendered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/RenderedTemplate"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "RenderTemplateRequest": {
        "type": "object",
        "required": ["templateId", "variables"],
        "properties": {
          "templateId": {
            "type": "string"
          },
          "variables": {
            "type": "object"
          },
          "language": {
            "type": "string",
            "default": "en"
          },
          "personalizeFor": {
            "type": "object",
            "properties": {
              "userId": {
                "type": "string"
              },
              "deviceType": {
                "type": "string",
                "enum": ["mobile", "desktop", "tablet"]
              },
              "preferences": {
                "type": "object"
              }
            }
          },
          "preview": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "RenderedTemplate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string"
                },
                "action": {
                  "type": "string"
                },
                "priority": {
                  "type": "string"
                }
              }
            }
          },
          "language": {
            "type": "string"
          },
          "personalizedFor": {
            "type": "object"
          },
          "variables": {
            "type": "object"
          },
          "renderedAt": {
            "type": "string",
            "format": "date-time"
          },
          "wordCount": {
            "type": "integer"
          },
          "estimatedReadTime": {
            "type": "integer",
            "description": "Estimated reading time in minutes"
          }
        }
      },
...
    },
...
  }
}